requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
pymupdf
chromadb
groq
langchain-groq